rename notifier to slack_channel_notifier
